Finished Scrum-80, Scrum-56 and Scrum-83
The scrums indicated in the title are over.

The user registration is now correctly operational, as well as the logging. A component was implemented to modify the data of a registered user, if desired. The navbar has more user-friendly modifications, hiding logout and user modification, for when it is not logged, and in case it is logged, it hides the loggin and shows the user modification and logout.