Finished Scrum 48 and Scrum 77
The prePayment component was added and improvements were made to the validation service. Now a guest customer, when they want to register, will be able to do so normally, since the validation is carried out with customers associated with users, not as previously it was done against all customers. The created component checks if there is a logged-in user, if it is affirmative it redirects him to the payment, in case it is negative, it allows him to log in from this component, or buy as a guest.